+++ pages/liked.tsx
@@ -1,4 +1,4 @@
-import { DehydratedState } from 'react-query';
+import type { DehydratedState } from 'react-query';
 
 import { Api, configureQueryClient } from '@graasp/query-client';
 
@@ -6,7 +6,7 @@ import Wrapper from '../src/components/common/Wrapper';
 import MyList from '../src/components/pages/MyList';
 import { QUERY_CLIENT_OPTIONS } from '../src/config/queryClient';
 
-const MyListPage = ({
+const MyLikedItemsPage = ({
   dehydratedState,
 }: {
   dehydratedState: DehydratedState;
@@ -17,16 +17,15 @@ const MyListPage = ({
 );
 
 export async function getServerSideProps() {
-  const { queryClient, dehydrate } = configureQueryClient(QUERY_CLIENT_OPTIONS);
+  const { queryClient, dehydrate, axios } =
+    configureQueryClient(QUERY_CLIENT_OPTIONS);
 
   await queryClient.prefetchQuery(['items', 'collections', 'all'], () =>
-    Api.getAllPublishedItems({}, QUERY_CLIENT_OPTIONS).then((data) =>
-      JSON.parse(JSON.stringify(data)),
-    ),
+    Api.getAllPublishedItems({}, { ...QUERY_CLIENT_OPTIONS, axios }),
   );
 
   // Pass data to the page via props
   return { props: { dehydratedState: dehydrate(queryClient) } };
 }
 
-export default MyListPage;
+export default MyLikedItemsPage;