+++ src/components/collection/summary/SummaryHeader.tsx
@@ -1,27 +1,33 @@
 import dynamic from 'next/dynamic';
+import Link from 'next/link';
 
 import React, { useContext } from 'react';
 
-import { Favorite } from '@mui/icons-material';
+import { Favorite, Visibility } from '@mui/icons-material';
 import { Skeleton } from '@mui/lab';
 import {
+  Alert,
   Box,
   Chip,
   Container,
   Divider,
+  Snackbar,
   Stack,
   Tooltip,
   Typography,
 } from '@mui/material';
 
-import { ThumbnailSize } from '@graasp/sdk';
-import { ItemLikeRecord, ItemRecord } from '@graasp/sdk/frontend';
+import { DiscriminatedItem, ThumbnailSize } from '@graasp/sdk';
 
-import { GRAASP_COLOR } from '../../../config/constants';
+import { UrlSearch } from '../../../config/constants';
+import { useLibraryTranslation } from '../../../config/i18n';
+import { ALL_COLLECTIONS_ROUTE } from '../../../config/routes';
 import {
   ITEM_SUMMARY_TITLE_ID,
+  LIKE_COLLECTION_NOT_LOGGED_ID,
   SUMMARY_TAGS_CONTAINER_ID,
 } from '../../../config/selectors';
+import LIBRARY from '../../../langs/constants';
 import { QueryClientContext } from '../../QueryClientContext';
 import CardMedia from '../../common/CardMediaComponent';
 import { StyledCard } from '../../common/StyledCard';
@@ -38,11 +44,12 @@ const { LikeButton } = {
 };
 
 type SummaryHeaderProps = {
-  collection?: ItemRecord;
+  collection?: DiscriminatedItem;
   isLoading: boolean;
   truncatedName: string;
-  tags: Immutable.List<string> | undefined;
+  tags?: string[];
   isLogged: boolean;
+  totalViews: number;
 };
 
 const SummaryHeader: React.FC<SummaryHeaderProps> = ({
@@ -51,24 +58,47 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
   isLoading,
   truncatedName,
   tags,
+  totalViews,
 }) => {
+  const { t } = useLibraryTranslation();
+
+  const [open, setOpen] = React.useState(false);
+
   const { hooks, mutations } = useContext(QueryClientContext);
 
   const { data: member } = hooks.useCurrentMember();
   const { data: likedItems } = hooks.useLikesForMember(member?.id);
   const { data: itemLikesForItem } = hooks.useLikesForItem(collection?.id);
-  const likes = itemLikesForItem?.size;
+  const likes = itemLikesForItem?.length;
 
   const { mutate: postItemLike } = mutations.usePostItemLike();
   const { mutate: deleteItemLike } = mutations.useDeleteItemLike();
 
   const likeEntry = likedItems?.find(
-    (itemLike: ItemLikeRecord) => itemLike?.item.id === collection?.id,
+    (itemLike) => itemLike?.item.id === collection?.id,
   );
 
+  const openLoginSnackbarMessage = () => {
+    setOpen(true);
+  };
+
+  const handleCloseSnackBarMessage = (
+    event: React.SyntheticEvent | Event,
+    reason?: string,
+  ) => {
+    if (reason === 'clickaway') {
+      return;
+    }
+
+    setOpen(false);
+  };
   const handleLike = () => {
-    if (!collection?.id || !member?.id) {
+    if (!collection?.id) {
       console.error('unable to like an item which id is undefined');
+      return;
+    }
+    if (!member?.id) {
+      openLoginSnackbarMessage();
       return;
     }
     postItemLike({
@@ -78,8 +108,12 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
   };
 
   const handleUnlike = () => {
-    if (!collection?.id || !member?.id) {
+    if (!collection?.id) {
       console.error('unable to unlike an item which id is undefined');
+      return;
+    }
+    if (!member?.id) {
+      openLoginSnackbarMessage();
       return;
     }
     deleteItemLike({
@@ -135,19 +169,17 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
               >
                 {truncatedName}
               </Typography>
-              {member && member.id && (
-                <LikeButton
-                  ariaLabel="like"
-                  color="primary"
-                  isLiked={Boolean(likeEntry)}
-                  handleLike={handleLike}
-                  handleUnlike={handleUnlike}
-                />
-              )}
+              <LikeButton
+                ariaLabel="like"
+                color="primary"
+                isLiked={Boolean(likeEntry)}
+                handleLike={handleLike}
+                handleUnlike={handleUnlike}
+              />
             </Stack>
             <SummaryActionButtons item={collection} isLogged={isLogged} />
           </Stack>
-          {tags && tags.size && (
+          {tags && tags.length && (
             <Stack
               id={SUMMARY_TAGS_CONTAINER_ID}
               direction="row"
@@ -155,7 +187,15 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
               spacing={1}
             >
               {tags.map((text) => (
-                <Chip key={text} label={text} />
+                <Chip
+                  component={Link}
+                  href={{
+                    pathname: ALL_COLLECTIONS_ROUTE,
+                    query: { [UrlSearch.KeywordSearch]: text },
+                  }}
+                  key={text}
+                  label={text}
+                />
               ))}
             </Stack>
           )}
@@ -182,30 +222,28 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
               divider={
                 <Divider
                   flexItem
-                  sx={{
-                    alignSelf: 'center',
-                    color: GRAASP_COLOR,
-                    fontWeight: 'bold',
-                  }}
-                >
-                  {String.fromCharCode(183)}
-                </Divider>
+                  variant="middle"
+                  orientation="vertical"
+                  sx={{ margin: 1 }}
+                />
               }
             >
-              {/* turn on again when endpoint exists */}
-              {/* <Tooltip title="Views" arrow placement="bottom">
-                <Stack direction="row" alignItems="center">
-                  <Typography
-                    fontWeight="bold"
-                    display="flex"
-                    alignItems="center"
-                    color="primary"
-                  >
-                    {views}
-                  </Typography>
-                  <Visibility color="primary" style={{ marginLeft: 5 }} />
-                </Stack>
-              </Tooltip> */}
+              {/* display only when there's a views */}
+              {totalViews ? (
+                <Tooltip title="Views" arrow placement="bottom">
+                  <Stack direction="row" alignItems="center">
+                    <Typography
+                      fontWeight="bold"
+                      display="flex"
+                      alignItems="center"
+                      color="primary"
+                    >
+                      {totalViews}
+                    </Typography>
+                    <Visibility color="primary" style={{ marginLeft: 5 }} />
+                  </Stack>
+                </Tooltip>
+              ) : null}
               <Tooltip title="Likes" arrow placement="bottom">
                 <Stack direction="row" alignItems="center">
                   <Typography
@@ -227,6 +265,20 @@ const SummaryHeader: React.FC<SummaryHeaderProps> = ({
           </Stack>
         </Stack>
       </Stack>
+      <Snackbar
+        anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
+        open={open}
+        autoHideDuration={6000}
+        onClose={handleCloseSnackBarMessage}
+      >
+        <Alert
+          id={LIKE_COLLECTION_NOT_LOGGED_ID}
+          onClose={handleCloseSnackBarMessage}
+          severity="error"
+        >
+          {t(LIBRARY.SIGNIN_MESSAGE)}
+        </Alert>
+      </Snackbar>
     </Container>
   );
 };